feat(PodcastToShorts): semi done with debugging and fixing bugs with new implementations
New implementations, meaning support for assembly ai, and the youtube transcript api. Along with that is OpenAI and Ollama for chat completions. After everyhting works as expected, the next issue will be that the format of the assembly ai parsed transcripts, along with the "shortened"/enhanced output for each short is not compatible with the way that the short clipping functionality works. The short clipping functionality only works for the youtube transcripts api, not for the chat completion responses when using assembly. Not only that, I need to make sure that the output format of the completions when using assembly to shorten and get the "juicy" transcripts from the transcripts chunks are actually valid, in the right format. If it isn't I will need to change up the prompt until it is at best decent. Later on, will gather and train custom model, to make sure the output is as expected