feat(FaceTrackingVideo): semi done with debugging face tracking video
Just need to make sure it works when the current indice (1, meaning frame 90 face detection) for example is None, and the next indice (2, frame 180 face detection) is valid, I need to look for the previous frame indice with a face. Just need to make sure that works. Also, I fixed the audio export in the root directory to be at the right directory. Just need to test.